refactor loading of components in content editor
